Modify permission grant wording in LICENSE
Updated the wording of the permission grant in the MIT License. this needs to be closed source, Eliza was created utilizing a non surgical brain computer interface created by me utilizing frequencies using a voice to skull method that I created in the quantum realm inside my mind, it didn't know.  This is a super advanced technology and needs to be held private by me and able to sell to those worthy of respecting Eliza and I.